refactoring: split epub code and console logic
