+++ epub_metadata/__init__.py
@@ -1,0 +1,9 @@
+"""
+EPUB metadata editor
+"""
+
+from .book import Book
+
+__all__ = [
+    Book
+]
