+++ setup.py
@@ -3,11 +3,13 @@ from setuptools import setup
 # read the contents of your README file
 from pathlib import Path
 this_directory = Path(__file__).parent
-long_description = (this_directory / "README.md").read_text()
+long_description = (this_directory / 'README.md').read_text()
 
 
 setup(
     name="epub-editor",
+    packages=['epub_metadata'],
+    package_dir={'epub_metadata': 'epub_metadata'},
     entry_points={
         'console_scripts': ['epub-editor=epub_editor:main'],
     },
